Add GazetteerTableDuplicate to ignore files
This new class will match on the name of files, but will do nothing when
asked to upload them.

This deals with duplicate files, such as the country .zip files provided
by the US NGA. They contain a comprehensive file and then several files
split out by feature type. The split out files need to be ignored.

It is still possible to manually force the type/table for a file upload
regardless of the name.